JWT Implemented, DJANGO_AUTH_TOKEN to archive
