+++ app/user/views.py
@@ -3,15 +3,9 @@ Views for the user API.
 """
 
 from rest_framework import generics, permissions
-from rest_framework.authtoken.views import ObtainAuthToken
-from rest_framework.settings import api_settings
 from rest_framework_simplejwt.authentication import JWTAuthentication
 from rest_framework_simplejwt.views import TokenObtainPairView
-from user.serializers import (
-    AuthTokenSerializer,
-    CustomTokenObtainPairSerializer,
-    UserSerializer,
-)
+from user.serializers import CustomTokenObtainPairSerializer, UserSerializer
 
 
 # NOTE: generics.CreateAPIView is Django REST Framework specified POST request handler.
@@ -25,20 +19,6 @@ class CreateUserView(generics.CreateAPIView):
     # 4. After validation, CreateAPIView calls serializer's save() method, which in turn calls the serializer's create() method
     # 5. Finally, the view returns an appropriate response
     serializer_class = UserSerializer
-
-
-# NOTE: ObtainAuthToken is a built-in view in Django REST Framework
-# That accepts POST requests to allow users to obtain an authentication token by submitting their username and password.
-# DOC: https://www.django-rest-framework.org/api-guide/authentication/
-class CreateTokenView(ObtainAuthToken):
-    """Create a new auth token for user."""
-
-    # NOTE: post() of ObtainAuthToken runs serializer.is_valid(), and it runs validate() inside
-    serializer_class = AuthTokenSerializer
-
-    # NOTE: This is a setting within DRF that specifies the default renderer classes that DRF should use
-    # if not otherwise specified in a view. By default, it usually includes JSONRenderer.
-    renderer_classes = api_settings.DEFAULT_RENDERER_CLASSES
 
 
 class ManageUserView(generics.RetrieveUpdateAPIView):
